Fix UTF-8 encoding better by using jsonlite for fromJSON. Still to do: switch to jsonlite::toJSON. Change results in VariableGroup@entities being class list and not character sometimes.

+++ pkg/R/categories.R
@@ -49,7 +49,7 @@ setMethod("values", "Categories", function (x) vapply(x, value, numeric(1)))
 
 ##' @rdname Categories
 ##' @export
-setMethod("ids", "Categories", function (x) vapply(x, id, numeric(1)))
+setMethod("ids", "Categories", function (x) vapply(x, id, integer(1)))
 ##' @export
 setMethod("ids", "list", function (x) sapply(x, id)) ## for summaries
